Merges 20x80 frame size where reordering disabled
    Finishes merging in reordering 20x80 (if reordering disabled) with reordering 4x16 (if reordering selected)
    Fixes HxtItem::addFrameDataToHistogramWithSum(frame) turning negative pixel(s) into negative bin(s) - fixes two cases of Next Frame Correction crashing
    Rearranges ImageProcessor's charged sharing, total spectrum initialisation enabling different sized histograms depending on reordering choice
    Ensures that Charge Sharing has a default value; cleans out some unused code.

+++ HxtProcessingLib/hxtitem.cpp
@@ -146,18 +146,18 @@ void HxtItem::addFrameDataToHistogramWithSum(double *frame)
    double *currentHistogram = &histogramPerPixel[0];
    long long *summed = &summedHistogram[0];
    double thisEnergy;
-   int bin;
-   int pixel;
+   uint64_t bin;
+   uint32_t pixel;
 
-   int frameSize = hxtV3Buffer.nRows * hxtV3Buffer.nCols;
-   for (int i = 0; i < frameSize; i++)
+   uint32_t frameSize = hxtV3Buffer.nRows * hxtV3Buffer.nCols;
+   for (uint32_t i = 0; i < frameSize; i++)
    {
       pixel = i;
       thisEnergy = frame[i];
 
-      if (thisEnergy == 0)
+      if (thisEnergy == 0.0)
           continue;
-      bin = (int)((thisEnergy / binWidth));
+      bin = (uint64_t)((thisEnergy / binWidth));
       if (bin <= nBins)
       {
          (*(currentHistogram + (pixel * nBins) + bin))++;
